Moving test to thier own directory
