+++ package.json
@@ -14,7 +14,7 @@
   "scripts": {
     "develop": "node dist/scripts/develop.js",
     "start": "ts-node-dev --inspect=8091 src/scripts/serve.ts",
-    "test": "mocha --timeout 10000 -r ts-node/register src/*.test.ts src/**/*.test.ts src/**/**/*.test.ts",
+    "test": "mocha --timeout 10000 -r ts-node/register test/*.test.ts",
     "lint": "eslint \"**/*.{js,ts}\"",
     "lint:fix": "eslint \"**/*.{js,ts}\" --quiet --fix",
     "prettier": "prettier --check \"**/*.{js,ts}\"",
